platform: mediatek: Add support for mt8196 platform
mt8196 plaform integrates a single-core HIFI5 DSP

The DSP does not have a DVFS scenario. When the system is in an active state,
the DSP clock is 800MHz(0.75v). When the system enters a low power scenario,
the DSP is in WFI state, and the clock is 26MHz.

The DSP core and DMA in DSP subsys both can access DRAM directly.

Signed-off-by: hailong.fan <hailong.fan@mediatek.com>